AI-Powered-Enhanced EHR Imaging & Documentation System/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
opencv-python-headless
pillow
numpy
pydantic
python-multipart       